Add the ability to force colorized compiler output
When CMake is configured to use the Ninja build
system, the compiler output to the terminal is not
colorized. This commit introduces a new CMake option
`enable_colorized_output` that forces colorized
outputs with the GNU and Clang compilers. It's
a simple cosmetic fix that shouldn't affect the
resulting builds otherwise.